+++ services/horizon/internal/actions_assets.go
@@ -36,7 +36,7 @@ func (action *AssetsAction) JSON() error {
 		action.loadParams,
 		action.loadRecords,
 		action.loadPage,
-		func() { hal.Render(action.W, action.Page) },
+		func() { hal.Render(action.W, action.Page, action.App.config.IsIndentedJSON) },
 	)
 	return action.Err
 }