+++ services/horizon/internal/actions_root.go
@@ -30,8 +30,9 @@ func (action *RootAction) JSON() error {
 		action.App.currentProtocolVersion,
 		action.App.coreSupportedProtocolVersion,
 		action.App.config.FriendbotURL,
+		action.App.config.ShouldPopulateHalCustomLinks,
 	)
 
-	hal.Render(action.W, res)
+	hal.Render(action.W, res, action.App.config.IsIndentedJSON)
 	return action.Err
 }