+++ services/horizon/internal/actions_trade.go
@@ -40,7 +40,7 @@ func (action *TradeIndexAction) JSON() error {
 		action.loadParams,
 		action.loadRecords,
 		action.loadPage,
-		func() { hal.Render(action.W, action.Page) },
+		func() { hal.Render(action.W, action.Page, action.App.config.IsIndentedJSON) },
 	)
 	return action.Err
 }
@@ -160,7 +160,7 @@ func (action *TradeAggregateIndexAction) JSON() error {
 		action.loadParams,
 		action.loadRecords,
 		action.loadPage,
-		func() { hal.Render(action.W, action.Page) },
+		func() { hal.Render(action.W, action.Page, action.App.config.IsIndentedJSON) },
 	)
 	return action.Err
 }