+++ services/horizon/internal/resourceadapter/transaction_success.go
@@ -10,14 +10,18 @@ import (
 )
 
 // Populate fills out the details
-func PopulateTransactionSuccess(ctx context.Context, dest *TransactionSuccess, result txsub.Result) {
+func PopulateTransactionSuccess(ctx context.Context, dest *TransactionSuccess, result txsub.Result, shouldPopulateHalCustomLinks bool) {
 	dest.Hash = result.Hash
 	dest.Ledger = result.LedgerSequence
 	dest.Env = result.EnvelopeXDR
 	dest.Result = result.ResultXDR
 	dest.Meta = result.ResultMetaXDR
 
-	lb := hal.LinkBuilder{httpx.BaseURL(ctx)}
-	dest.Links.Transaction = lb.Link("/transactions", result.Hash)
+	if shouldPopulateHalCustomLinks {
+		lb := hal.LinkBuilder{httpx.BaseURL(ctx)}
+		dest.Links = new(TransactionSuccessLinks)
+		dest.Links.Transaction = lb.LinkPtr("/transactions", result.Hash)
+	}
+
 	return
 }