+++ services/horizon/CHANGELOG.md
@@ -6,6 +6,11 @@ file.  This project adheres to [Semantic Versioning](http://semver.org/).
 As this project is pre 1.0, breaking changes may happen for minor version
 bumps.  A breaking change will get clearly notified in this log.
 
+## v0.18.2
+
+* Added configuration toggle to return JSON response with unindented JSON. This improves rendering and request rate performance.
+* Added configuration toggle to disable non-paginating HAL `"_links"` in JSON response. This improves rendering and request rate performance.
+
 ## v0.18.1
 
 * Fixed `/fee_stats` to correctly calculate ledger capacity in protocol v11.